+++ module/core/variadic_from/tests/inc/mod.rs
@@ -1,20 +1,18 @@
 use super::*;
 
 #[ cfg( all( feature = "type_variadic_from" ) ) ]
+mod from2_named_manual;
+#[ cfg( all( feature = "derive_variadic_from", feature = "type_variadic_from" ) ) ]
+mod from2_named_derive;
+
+// #[ cfg( all( feature = "derive_variadic_from", feature = "type_variadic_from" ) ) ]
+// mod from2_unnamed;
+// xxx : fix
+
+#[ cfg( all( feature = "type_variadic_from" ) ) ]
 mod from4_named_manual;
 #[ cfg( all( feature = "type_variadic_from" ) ) ]
 mod from4_unnamed_manual;
-
-// xxx
-// #[ cfg( all( feature = "derive_variadic_from", feature = "type_variadic_from" ) ) ]
-// mod auto_std_named_derive;
-
-#[ cfg( all( feature = "derive_variadic_from", feature = "type_variadic_from" ) ) ]
-mod auto_std_named_manual;
-
-// #[ cfg( all( feature = "derive_variadic_from", feature = "type_variadic_from" ) ) ]
-// mod variadic_from_derive;
-// xxx : fix
 
 #[ cfg( all( feature = "type_variadic_from" ) ) ]
 mod from4_beyond_named;
