+++ module/core/variadic_from/tests/inc/only_test/from2_named.rs
@@ -1,8 +1,8 @@
 #[ test ]
-fn variadic_from2_named()
+fn from2_named()
 {
 
-  //
+  // - from_2
 
   let got : StructNamedFields = from!( 13, 14 );
   let exp = StructNamedFields{ a : 13, b : 14 };
@@ -16,6 +16,8 @@ fn variadic_from2_named()
   let exp = StructNamedFields{ a : 13, b : 14 };
   a_id!( got, exp );
 
+  // - from_1
+
   let got : StructNamedFields = StructNamedFields::from_1( ( 13, 14 ) );
   let exp = StructNamedFields{ a : 13, b : 14 };
   a_id!( got, exp );
@@ -28,6 +30,8 @@ fn variadic_from2_named()
   let exp = StructNamedFields{ a : 13, b : 14 };
   a_id!( got, exp );
 
+  // - to
+
   let got : StructNamedFields = ( 13, 14 ).to();
   let exp = StructNamedFields{ a : 13, b : 14 };
   a_id!( got, exp );
@@ -36,6 +40,14 @@ fn variadic_from2_named()
   let exp = StructNamedFields{ a : 13, b : 14 };
   a_id!( got, exp );
 
-  //
+  // - std
+
+  let got : StructNamedFields = From::from( ( 13, 14 ) );
+  let exp = StructNamedFields{ a : 13, b : 14 };
+  a_id!( got, exp );
+
+  let got : StructNamedFields = ( 13, 14 ).into();
+  let exp = StructNamedFields{ a : 13, b : 14 };
+  a_id!( got, exp );
 
 }
